\name{ScatterDensity-package}
\alias{ScatterDensity-package}
\alias{ScatterDensity}
\docType{package}
\title{
\packageTitle{ScatterDensity}
}
\description{
\packageDescription{ScatterDensity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ScatterDensity}
\packageIndices{ScatterDensity}
}
\author{
\packageAuthor{ScatterDensity}

Maintainer: \packageMaintainer{ScatterDensity}
}

\keyword{ package }

\examples{
#Todo
}
